Migrate frameobject.h contents to cpython/frameobject.h
